Implemented contribution page
I implemented the contribution page to handle N number of contributors. I didn't abstract the Cards because I'm not sure that they will be reused. Can be extracted to a component if need be. The color pattern doesn't follow the mockup to a tee, but enforces that the 3 secondary colors are used in a row and that each row starts with a different color.